add functionality to update instrument
Instrument epochs are updated with the definition of new
throughput calibration methods. This does not affect the
analysis of the evndisp step for data files, but the epoch
string in VEvndispRunParameter is used in many workflows.

This changes allow to:
- update the VEvndispRunParameter epoch string
- get an updated instrument epoch using printRunParameter

Note that the implement depend on the default file with
all instrument epochs (VERITAS.Epochs.runparameter)